requirements.txt
flask
python-dotenv
requests
google-genai
markdown2
pytest
playwright
psycopg2-binary
